requirements.txt
 hydra-core==1.2.0
 numpy==1.22.4 
 wandb==0.13.1 
 easydict==1.9 
 transformers==4.21.1
 opencv-python==4.6.0.66 
 robomimic==0.2.0 
 einops==0.4.1
 thop==0.1.1-2209072238
 bddl==1.0.1
 future==0.18.2 
 matplotlib==3.5.3 
 cloudpickle==2.1.0
 gym==0.25.2
 timm==1.0.15
 torch==2.4.1+cu118
 torchvision==0.19.1+cu118
 torchtext==2.4.1+cu118
 robosuite==1.4.1
 scikit-learn==1.3.2